refactor(examples): simplify and rename memory-backed deploy script
